+++ okama/api/api_methods.py
@@ -1,5 +1,3 @@
-from typing import Union  # noqa: I001
-
 import pandas as pd
 import requests
 from urllib3.util.retry import Retry
@@ -36,8 +34,8 @@ class API:
         cls,
         endpoint: str = endpoint_ror,
         symbol: str = settings.default_ticker,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
         period: str = "d",
     ) -> str:
         session = requests.session()
@@ -80,8 +78,8 @@ class API:
     def get_ror(
         cls,
         symbol: str = settings.default_ticker,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
         period: str = "m",
     ):
         return cls.connect(
@@ -96,8 +94,8 @@ class API:
     def get_adjusted_close(
         cls,
         symbol: str = settings.default_ticker,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
         period: str = "m",
     ):
         return cls.connect(
@@ -112,8 +110,8 @@ class API:
     def get_close(
         cls,
         symbol: str = settings.default_ticker,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
         period: str = "m",
     ):
         return cls.connect(
@@ -128,8 +126,8 @@ class API:
     def get_dividends(
         cls,
         symbol: str = settings.default_ticker,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
     ):
         return cls.connect(
             endpoint=cls.endpoint_dividends,
@@ -142,8 +140,8 @@ class API:
     def get_nav(
         cls,
         symbol: str = settings.default_ticker,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
         period: str = "m",
     ):
         return cls.connect(
@@ -158,8 +156,8 @@ class API:
     def get_macro(
         cls,
         symbol: str = settings.default_ticker,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
         period: str = "m",
     ):
         """