+++ okama/macro.py
@@ -1,5 +1,4 @@
 from abc import ABC  # noqa: I001
-from typing import Union, Tuple  # noqa: UP035
 
 import numpy as np
 import pandas as pd
@@ -30,8 +29,8 @@ class MacroABC(ABC):  # noqa: B024
     def __init__(
         self,
         symbol: str,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
     ):
         self.symbol: str = symbol
         self._check_namespace()
@@ -146,7 +145,7 @@ class MacroABC(ABC):  # noqa: B024
         self._values_monthly[pd.Period(date, freq="M")] = value
         self._set_first_last_dates()
 
-    def describe(self, years: Tuple[int, ...] = (1, 5, 10)) -> pd.DataFrame:  # noqa: UP006
+    def describe(self, years: tuple[int, ...] = (1, 5, 10)) -> pd.DataFrame:
         """
         Generate descriptive statistics for YTD and given periods.
         Statistics includes:
@@ -260,8 +259,8 @@ class Inflation(MacroABC):
     def __init__(
         self,
         symbol: str = settings.default_macro_inflation,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
     ):
         super().__init__(
             symbol,
@@ -375,7 +374,7 @@ class Inflation(MacroABC):
         x = x.dropna()
         return x
 
-    def describe(self, years: Tuple[int, ...] = (1, 5, 10)) -> pd.DataFrame:  # noqa: UP006
+    def describe(self, years: tuple[int, ...] = (1, 5, 10)) -> pd.DataFrame:
         """
         Generate descriptive inflation statistics for YTD and a given list of periods.
         Statistics includes:
@@ -517,8 +516,8 @@ class Rate(MacroABC):
     def __init__(
         self,
         symbol: str = settings.default_macro_rate,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
     ):
         super().__init__(
             symbol,
@@ -564,8 +563,8 @@ class Indicator(MacroABC):
     def __init__(
         self,
         symbol: str = settings.default_macro_indicator,
-        first_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
-        last_date: Union[str, pd.Timestamp, None] = None,  # noqa: UP007
+        first_date: str | pd.Timestamp | None = None,
+        last_date: str | pd.Timestamp | None = None,
     ):
         super().__init__(
             symbol,